Finish 04_removeFromArray exercise
Write a function that takes an array and some other arguments then
removes the other arguments from that array.

Test it with the removeFromArray.spec.js file, first with the first
test of the file and then enabling the other 7 tests by removing the
".skip" from the "test.skip()" functions.