Finish 07_tempConversion exercise
Write two functions that convert temperatures from Fahrenheit to
Celsius, and vice versa, rounded to one decimal place.

Test it with the tempConversion.spec.js file, first with the first
tests of the file of each function and then enabling the other tests
functions by removing the ".skip" from the "test.skip()" functions.